feat: move UserFeedViewModel and StringFeedViewModel to commons
Move UserFeedViewModel, UserFeedState, StringFeedViewModel, and
StringFeedState to commons module for KMP iOS migration.

- Create commons UserFeedViewModel with ICacheProvider parameter
- Create commons StringFeedViewModel with ICacheProvider parameter
- Leave Android wrappers (AndroidUserFeedViewModel, AndroidStringFeedViewModel)
  that auto-provide LocalCache
- Update subclasses to use Android wrappers
- Add explicit imports in View files for commons types
- Delete empty state files (now co-located with their ViewModels)
