+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/settings/dal/HiddenWordsFeedViewModel.kt
@@ -23,11 +23,11 @@ package com.vitorpamplona.amethyst.ui.screen.loggedIn.settings.dal
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.vitorpamplona.amethyst.model.Account
-import com.vitorpamplona.amethyst.ui.screen.loggedIn.settings.StringFeedViewModel
+import com.vitorpamplona.amethyst.ui.screen.loggedIn.settings.AndroidStringFeedViewModel
 
 class HiddenWordsFeedViewModel(
     val account: Account,
-) : StringFeedViewModel(
+) : AndroidStringFeedViewModel(
         HiddenWordsFeedFilter(account),
     ) {
     class Factory(
